refactor(plans): move PlanTree to own folder + extract PlanTreeDialogs
Relocates PlanTree.tsx and its six dialog components from
src/components/SkillPlans/ into a dedicated src/components/PlanTree/
folder, with PlanTree.tsx renamed to index.tsx.

Adds PlanTreeDialogs.tsx as a self-contained dialog host: it reads the
active dialog from the Zustand store, owns the delete mutation +
post-create navigation, and reads the current planId straight from the
router. PlanTree no longer carries any dialog state, mutation, or JSX.

Clears the react-doctor no-giant-component warning on PlanTree.
